server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
youtube-transcript-api>=0.6.0
langchain>=0.1.0
langchain-community>=0.0.38
langchain-core>=0.1.0
langchain-text-splitters>=0.1.0
langchain-ollama>=0.0.1
faiss-cpu>=1.13.0
python-dotenv>=1.0.0
# Removed HuggingFace specific libs
# huggingface-hub>=0.20.0
# sentence-transformers>=2.2.0
# torch>=2.0.0
# transformers>=4.30.0
scikit-learn>=1.3.0
numpy>=1.24.0
protobuf>=3.20.0
sentencepiece>=0.1.99
requests>=2.31.0
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.0
google-api-python-client>=2.100.0